fix unicode strings
